partial linting for data scripts
